refactor: Remove duplicated code in viewmodels
